umcu/requirements.txt
Django==1.9.6
django-localflavor==1.3
MySQL-python==1.2.5
